refactor: old BottomSlider removed
